+++ src/HasFootLength.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Bakame\Shoes;
+
+interface HasFootLength
+{
+    public Length $footLength {
+        get;
+    }
+}